Renamed Base to ContextBase and moved it into context
