+++ snek_sploit/lib/context.py
@@ -1,5 +1,6 @@
 import requests
 import msgpack
+from abc import ABC
 from typing import Union, List, Dict
 
 from snek_sploit.util import constants
@@ -82,3 +83,31 @@ class Context:
             raise Exception(response)
 
         return response
+
+
+class ContextBase(ABC):
+    def __init__(self, context: Context):
+        self._context = context
+
+    def _decode(self, to_decode: Union[dict, list, bytes, str, int]) -> Union[dict, list, str, int]:
+        """
+        Decode everything into the correct type and decode bytes.
+        :param to_decode: Input you want to decode
+        :return: Decoded input
+        """
+        if isinstance(to_decode, dict):
+            decoded = {}
+            for key, value in to_decode.items():
+                decoded[self._decode(key)] = self._decode(value)
+        elif isinstance(to_decode, list):
+            decoded = [self._decode(each) for each in to_decode]
+        elif isinstance(to_decode, bytes):
+            decoded = to_decode.decode()
+        else:
+            decoded = to_decode
+
+        # TODO: also try to parse a datetime? Not sure if there is a united format in MSF
+        if isinstance(decoded, str) and decoded.isdigit():
+            decoded = int(decoded)
+
+        return decoded
