+++ snek_sploit/lib/groups/console.py
@@ -1,8 +1,8 @@
-from snek_sploit.lib.base import Base, Context
+from snek_sploit.lib.context import ContextBase, Context
 from snek_sploit.lib.rpc import RPCConsole, ConsoleInfo
 
 
-class Console(Base):
+class Console(ContextBase):
     def __init__(self, context: Context):
         super().__init__(context)
         self.rpc = RPCConsole(context)
