+++ snek_sploit/lib/groups/core.py
@@ -1,8 +1,8 @@
-from snek_sploit.lib.base import Base, Context
+from snek_sploit.lib.context import ContextBase, Context
 from snek_sploit.lib.rpc import RPCCore
 
 
-class Core(Base):
+class Core(ContextBase):
     def __init__(self, context: Context):
         super().__init__(context)
         self.rpc = RPCCore(context)
