+++ snek_sploit/lib/groups/db.py
@@ -1,8 +1,8 @@
-from snek_sploit.lib.base import Base, Context
+from snek_sploit.lib.context import ContextBase, Context
 from snek_sploit.lib.rpc import RPCDB
 
 
-class DB(Base):
+class DB(ContextBase):
     def __init__(self, context: Context):
         super().__init__(context)
         self.rpc = RPCDB(context)
