+++ snek_sploit/lib/groups/health.py
@@ -1,8 +1,8 @@
-from snek_sploit.lib.base import Base, Context
+from snek_sploit.lib.context import ContextBase, Context
 from snek_sploit.lib.rpc import RPCHealth
 
 
-class Health(Base):
+class Health(ContextBase):
     def __init__(self, context: Context):
         super().__init__(context)
         self.rpc = RPCHealth(context)
