+++ snek_sploit/lib/groups/module.py
@@ -1,8 +1,8 @@
-from snek_sploit.lib.base import Base, Context
+from snek_sploit.lib.context import ContextBase, Context
 from snek_sploit.lib.rpc import RPCModule
 
 
-class Module(Base):
+class Module(ContextBase):
     def __init__(self, context: Context):
         super().__init__(context)
         self.rpc = RPCModule(context)
