+++ snek_sploit/lib/groups/plugin.py
@@ -1,8 +1,8 @@
-from snek_sploit.lib.base import Base, Context
+from snek_sploit.lib.context import ContextBase, Context
 from snek_sploit.lib.rpc import RPCPlugin
 
 
-class Plugin(Base):
+class Plugin(ContextBase):
     def __init__(self, context: Context):
         super().__init__(context)
         self.rpc = RPCPlugin(context)
