+++ snek_sploit/lib/rpc/health.py
@@ -1,8 +1,8 @@
-from snek_sploit.lib.base import Base
+from snek_sploit.lib.context import ContextBase
 from snek_sploit.util import constants
 
 
-class RPCHealth(Base):
+class RPCHealth(ContextBase):
     """
     https://docs.metasploit.com/api/Msf/RPC/RPC_Health.html
     """
